backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
psycopg2-binary
python-dotenv
bcrypt
gunicorn
sqlalchemy